requirements.txt
pyserial
requests
Flask
Flask-SocketIO
vpype
svgwrite